requirement.txt
scrapy
beautifulsoup4
transformers==4.44.0
huggingface_hub==0.24.5
nltk==3.8.1
pandas
tensorflow
torch
tf-keras
gradio == 4.36.1
matplotlib
seaborn
spacy
pyvis == 0.3.2
scikit-learn
datasets
evaluate == 0.4.2
python-dotenv